Download symbols in lldb/sos and for portable PDBs on Windows/dbgeng.
Add symbol store support.

Separate the coreclr hosting code.

Use dlopen/dlsym on Linux to host coreclr.

Remove _EFN_ functions and VerifyStackTrace from xplat builds of SOS.

Add native projects for easier Strike development.

Fix packaging and yaml for additional symstore assemblies.

Don't pack dotnet-analyze and dotnet-collect because they currently break the official job.
