+++ src/SOS/lldbplugin/lldbplugin.vcxproj.filters
@@ -1,0 +1,23 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <ClCompile Include="services.cpp" />
+    <ClCompile Include="setclrpathcommand.cpp" />
+    <ClCompile Include="setsostidcommand.cpp" />
+    <ClCompile Include="soscommand.cpp" />
+    <ClCompile Include="sosplugin.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="mstypes.h" />
+    <ClInclude Include="services.h" />
+    <ClInclude Include="sosplugin.h" />
+    <ClInclude Include="inc\lldbservices.h">
+      <Filter>inc</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <Filter Include="inc">
+      <UniqueIdentifier>{32c8e536-f88f-4485-a168-9f9e9fab5822}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+</Project>
