+++ src/Tools/dotnet-analyze/dotnet-analyze.csproj
@@ -8,8 +8,10 @@
 
     <RootNamespace>Microsoft.Diagnostics.Tools.Analyze</RootNamespace>
 
+    <!-- Don't pack until ship engineering is done. Currently causing the official job to fail.
     <IsPackable>true</IsPackable>
     <PackAsTool>true</PackAsTool>
+    -->
   </PropertyGroup>
 
   <ItemGroup>
