+++ src/Tools/dotnet-collect/dotnet-collect.csproj
@@ -9,8 +9,10 @@
     <RootNamespace>Microsoft.Diagnostics.Tools.Collect</RootNamespace>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
 
+    <!-- Don't pack until ship engineering is done. Currently causing the official job to fail.
     <IsPackable>true</IsPackable>
     <PackAsTool>true</PackAsTool>
+    -->
   </PropertyGroup>
 
   <ItemGroup>
