+++ src/inc/arrayholder.h
@@ -2,6 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+#ifndef __ARRAYHOLDER_H__
+#define __ARRAYHOLDER_H__
+
 template <class T>
 class ArrayHolder    
 {
@@ -78,3 +81,5 @@ private:
 private:
     T *m_ptr;
 };
+
+#endif // __ARRAYHOLDER_H__
